Add support for lmdeploy. Merge the two generation scripts into one.

+++ generate_inference_shell_scripts/generate_infer_compare_shell.py
@@ -5,7 +5,9 @@ output_dir_compare = "/mnt/pfs/zitao_team/fangzhensheng/MathEval/all_compare_res
 log_dir="/mnt/pfs/zitao_team/fangzhensheng/MathEval/all_infer_compare_logs"
 where_to_save_shell = "/mnt/pfs/zitao_team/fangzhensheng/MathEval/all_infer_compare_shell"
 
-def generate_infer_compare_shell(all_data_info, root_model_info):
+def generate_infer_compare_shell(all_data_info, root_model_info, args):
+    accelerator = args.accelerator
+    mode = args.mode
     date = datetime.now().strftime("%Y_%m%d")
     script_file_names = []
     
@@ -40,6 +42,7 @@ def generate_infer_compare_shell(all_data_info, root_model_info):
             script_file.write(f"model_path={model_path}\n")
             script_file.write(f"model_name={model_name}\n")
             script_file.write(f"device_num={device_num}\n")
+            script_file.write(f"accelerator={accelerator}\n")
             log_file = os.path.join(log_dir_model_data, model_name + ".log")
             script_file.write(f"log_file={log_file}\n")
             
@@ -50,6 +53,7 @@ def generate_infer_compare_shell(all_data_info, root_model_info):
 
             script_file.write('command_infer="CUDA_VISIBLE_DEVICES=$gpu_ids nohup python3 -u /mnt/pfs/zitao_team/fangzhensheng/MathEval/inference_matheval.py')
             script_file.write(f" --model_name $model_name --model_path $model_path --device_num $device_num")
+            script_file.write(' --accelerator $accelerator')
             script_file.write(f' --output_dir $output_dir_infer --max_new_tokens 512 > $log_file 2>&1 &"\n\n')
 
             script_file.write('eval "$command_infer"\n')
@@ -62,29 +66,30 @@ def generate_infer_compare_shell(all_data_info, root_model_info):
             script_file.write("pkill -f inference_matheval.py\n\n")
 
             script_file.write('echo "Inference completed and related processes killed."\n\n')
+            
+            if mode == 'eval':
+                # Comparison variables
+                input_dir_model = os.path.join(output_dir_infer, date, model_name)
 
-            # Comparison variables
-            input_dir_model = os.path.join(output_dir_infer, date, model_name)
+                script_file.write(f'output_dir_compare="{output_dir_model_data_compare}"\n')
+                script_file.write(f'input_dir="{input_dir_model}"\n')
 
-            script_file.write(f'output_dir_compare="{output_dir_model_data_compare}"\n')
-            script_file.write(f'input_dir="{input_dir_model}"\n')
+                script_file.write(f'mkdir -p "$output_dir_compare"\n')
+                script_file.write(f'pids=() # Array to store the PIDs of the background processes\n\n')
+                script_file.write('command_compare="CUDA_VISIBLE_DEVICES=$gpu_ids nohup python3 -u /mnt/pfs/zitao_team/fangzhensheng/MathEval/compare_with_local_model.py')
+                script_file.write(f' --device_num $device_num')
+                script_file.write(f' --input_dir $input_dir --output_dir $output_dir_compare >> $log_file 2>&1 &"\n\n')
 
-            script_file.write(f'mkdir -p "$output_dir_compare"\n')
-            script_file.write(f'pids=() # Array to store the PIDs of the background processes\n\n')
-            script_file.write('command_compare="CUDA_VISIBLE_DEVICES=$gpu_ids nohup python3 -u /mnt/pfs/zitao_team/fangzhensheng/MathEval/compare_with_local_model.py')
-            script_file.write(f' --device_num $device_num')
-            script_file.write(f' --input_dir $input_dir --output_dir $output_dir_compare >> $log_file 2>&1 &"\n\n')
+                script_file.write('eval "$command_compare"\n')
+                script_file.write('pids+=($!) # Store the PID of the last command run in the background\n\n')
+                script_file.write('for pid in "${pids[@]}"; do\n')
+                script_file.write('    wait "$pid"\n')
+                script_file.write("done\n\n")
 
-            script_file.write('eval "$command_compare"\n')
-            script_file.write('pids+=($!) # Store the PID of the last command run in the background\n\n')
-            script_file.write('for pid in "${pids[@]}"; do\n')
-            script_file.write('    wait "$pid"\n')
-            script_file.write("done\n\n")
+                script_file.write("# Kill all processes related to evaluating the model no api\n")
+                script_file.write("pkill -f compare_with_local_model.py\n\n")
 
-            script_file.write("# Kill all processes related to evaluating the model no api\n")
-            script_file.write("pkill -f compare_with_local_model.py\n\n")
-
-            script_file.write('echo "Comparison completed and related processes killed."\n')
+                script_file.write('echo "Comparison completed and related processes killed."\n')
 
         # Make the script executable
         os.chmod(script_file_name, 0o755)
